# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sheetwise API
  version: 2.0.0
  description: Private spreadsheet analytics. One shared internal workspace; all analytical
    database paths remain server-side. Schemas and examples are defined in components.
servers:
- url: http://127.0.0.1:8000
  description: Local development; use your HTTPS internal origin for deployment.
security:
- WorkspaceSession: []
paths:
  /api/status:
    get:
      operationId: getWorkspaceStatus
      summary: Read sign-in and AI configuration status.
      description: Read sign-in and AI configuration status. All records belong to
        the single authenticated internal workspace.
      responses:
        '200':
          description: Request completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceStatus'
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /api/session:
    post:
      operationId: createWorkspaceSession
      summary: Create an HttpOnly workspace session.
      description: Create an HttpOnly workspace session. All records belong to the
        single authenticated internal workspace.
      responses:
        '200':
          description: Request completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResult'
          headers:
            Set-Cookie:
              description: HttpOnly, SameSite=Strict session cookie; Secure when COOKIE_SECURE=true.
              schema:
                $ref: '#/components/schemas/Text'
              example: sheetwise_session=(signed session); HttpOnly; SameSite=Strict;
                Path=/
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
    delete:
      operationId: deleteWorkspaceSession
      summary: Clear this browser’s workspace session cookie.
      description: Clear this browser’s workspace session cookie. All records belong
        to the single authenticated internal workspace.
      responses:
        '200':
          description: Request completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResult'
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /api/datasets:
    get:
      operationId: listDatasets
      summary: List uploaded datasets.
      description: List uploaded datasets. All records belong to the single authenticated
        internal workspace.
      responses:
        '200':
          description: Request completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: uploadDataset
      summary: Import and profile a CSV or XLSX file.
      description: Import and profile a CSV or XLSX file. All records belong to the
        single authenticated internal workspace.
      responses:
        '201':
          description: Created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadRequest'
  /api/imports:
    post:
      operationId: importCloudUpload
      summary: Import a completed private cloud upload.
      description: Import a completed private cloud upload. All records belong to
        the single authenticated internal workspace.
      responses:
        '201':
          description: Created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudImportRequest'
...
  /api/openapi.yaml:
    get:
      operationId: getApiSpecification
      summary: Download this API contract.
      description: Download this API contract. All records belong to the single authenticated
        internal workspace.
      responses:
        '200':
          description: Request completed.
          content:
            application/yaml:
              schema:
                $ref: '#/components/schemas/Text'
        '400':
          description: Bad Request — invalid fields, unsupported file or rejected
            query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — a valid workspace session is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — untrusted host or cross-origin write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found — dataset or answer does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large — request or file exceeds its limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests — import or analysis capacity is busy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Error — the request could not be completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable — AI or upload storage configuration, credits
            or provider availability prevents the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    WorkspaceSession:
      type: apiKey
      in: cookie
      name: sheetwise_session
      description: Signed workspace session. Loopback-only local mode can omit this
        cookie.
  schemas:
...
    WorkspaceStatus:
      type: object
      description: Workspace availability. aiConfigured reports key presence, not
        account credit availability.
      properties:
        authenticated:
          $ref: '#/components/schemas/Boolean'
        passwordRequired:
          $ref: '#/components/schemas/Boolean'
        aiConfigured:
          $ref: '#/components/schemas/Boolean'
        model:
          $ref: '#/components/schemas/ShortText'
        maxUploadBytes:
          $ref: '#/components/schemas/Count'
        localMode:
          $ref: '#/components/schemas/Boolean'
        directUploads:
          $ref: '#/components/schemas/Boolean'
      required:
      - authenticated
      - passwordRequired
      - aiConfigured
      - model
      - maxUploadBytes
      - localMode
      - directUploads
      additionalProperties: false
      example:
        authenticated: true
        passwordRequired: false
        aiConfigured: true
        model: gpt-5.6-sol
        maxUploadBytes: 104857600
        localMode: true
        directUploads: false
    SessionResult:
      type: object
      description: Whether the workspace session is authenticated.
      properties:
        authenticated:
          $ref: '#/components/schemas/Boolean'
      required:
      - authenticated
      additionalProperties: false
      example:
        authenticated: true
...
    ErrorResponse:
      type: object
      description: Error envelope used by all API operations.
      properties:
        error:
          $ref: '#/components/schemas/Error'
      required:
      - error
      additionalProperties: false
      example:
        error: *id022
...